reorg STATS.h5 writer modules and functions
