+++ src/nisarqa/parameters/constants/globals.py
@@ -142,53 +142,6 @@ def get_possible_pols(product_type):
 
 PRODUCT_SPECS_PATH = Path(__file__).parent.parent / "product_specs"
 
-
-# Directory Structure and Paths in QA STATS.h5 file
-STATS_H5_BASE_GROUP = "/science/%sSAR"
-STATS_H5_IDENTIFICATION_GROUP = STATS_H5_BASE_GROUP + "/identification"
-processing_group = "/processing"
-data_group = "/data"
-
-# QA Directory Structure and Paths in STATS.h5 file
-STATS_H5_QA_STATS_H5_BASE_GROUP = STATS_H5_BASE_GROUP + "/QA"
-STATS_H5_QA_PROCESSING_GROUP = (
-    STATS_H5_QA_STATS_H5_BASE_GROUP + processing_group
-)
-STATS_H5_QA_DATA_GROUP = STATS_H5_QA_STATS_H5_BASE_GROUP + data_group
-# Frequency group. Note: There are two '%s' here. The first is for the band,
-# the second for the frequency.
-# Example end result: '/science/%s/QA/data/frequency%s'
-STATS_H5_QA_FREQ_GROUP = (
-    STATS_H5_QA_DATA_GROUP + "/frequency%s"
-)  # Two '%s' here!
-
-# RFI Group
-STATS_H5_RFI_BASE_GROUP = STATS_H5_BASE_GROUP + "/RFI"
-STATS_H5_RFI_DATA_GROUP = STATS_H5_RFI_BASE_GROUP + data_group
-
-# CalTools
-STATS_H5_ABSCAL_STATS_H5_BASE_GROUP = (
-    STATS_H5_BASE_GROUP + "/absoluteRadiometricCalibration"
-)
-STATS_H5_ABSCAL_PROCESSING_GROUP = (
-    STATS_H5_ABSCAL_STATS_H5_BASE_GROUP + processing_group
-)
-STATS_H5_ABSCAL_DATA_GROUP = STATS_H5_ABSCAL_STATS_H5_BASE_GROUP + data_group
-
-STATS_H5_PTA_STATS_H5_BASE_GROUP = STATS_H5_BASE_GROUP + "/pointTargetAnalyzer"
-STATS_H5_PTA_PROCESSING_GROUP = (
-    STATS_H5_PTA_STATS_H5_BASE_GROUP + processing_group
-)
-STATS_H5_PTA_DATA_GROUP = STATS_H5_PTA_STATS_H5_BASE_GROUP + data_group
-
-STATS_H5_NEB_STATS_H5_BASE_GROUP = (
-    STATS_H5_BASE_GROUP + "/noiseEquivalentBackscatter"
-)
-STATS_H5_NEB_PROCESSING_GROUP = (
-    STATS_H5_NEB_STATS_H5_BASE_GROUP + processing_group
-)
-STATS_H5_NEB_DATA_GROUP = STATS_H5_NEB_STATS_H5_BASE_GROUP + data_group
-
 complex32 = np.dtype([("r", np.float16), ("i", np.float16)])
 
 # The are global constants and not functions nor classes,
@@ -213,22 +166,5 @@ __all__ = [
     "GCOV_DIAG_POLS",
     "GCOV_OFF_DIAG_POLS",
     "SEABORN_COLORBLIND",
-    "STATS_H5_BASE_GROUP",
-    "STATS_H5_IDENTIFICATION_GROUP",
-    "STATS_H5_QA_STATS_H5_BASE_GROUP",
-    "STATS_H5_QA_PROCESSING_GROUP",
-    "STATS_H5_QA_DATA_GROUP",
-    "STATS_H5_QA_FREQ_GROUP",
-    "STATS_H5_RFI_BASE_GROUP",
-    "STATS_H5_RFI_DATA_GROUP",
-    "STATS_H5_ABSCAL_STATS_H5_BASE_GROUP",
-    "STATS_H5_ABSCAL_PROCESSING_GROUP",
-    "STATS_H5_ABSCAL_DATA_GROUP",
-    "STATS_H5_PTA_STATS_H5_BASE_GROUP",
-    "STATS_H5_PTA_PROCESSING_GROUP",
-    "STATS_H5_PTA_DATA_GROUP",
-    "STATS_H5_NEB_STATS_H5_BASE_GROUP",
-    "STATS_H5_NEB_PROCESSING_GROUP",
-    "STATS_H5_NEB_DATA_GROUP",
     "complex32",
 ]
