+++ src/nisarqa/products/gcov.py
@@ -140,13 +140,13 @@ def verify_gcov(
             )
             log.info(f"QA Processing Parameters saved to {stats_file}")
 
-            nisarqa.rslc.copy_identification_group_to_stats_h5(
+            nisarqa.copy_identification_group_to_stats_h5(
                 product=product, stats_h5=stats_h5
             )
             log.info(f"Input file Identification group copied to {stats_file}")
 
             # Save frequency/polarization info from `pols` to stats file
-            nisarqa.rslc.save_nisar_freq_metadata_to_h5(
+            nisarqa.save_nisar_freq_metadata_to_h5(
                 product=product, stats_h5=stats_h5
             )
