+++ src/nisarqa/products/gslc.py
@@ -113,7 +113,7 @@ def verify_gslc(
             )
             log.info(f"QA Processing Parameters saved to {stats_file}")
 
-            nisarqa.rslc.copy_identification_group_to_stats_h5(
+            nisarqa.copy_identification_group_to_stats_h5(
                 product=product, stats_h5=stats_h5
             )
             log.info(f"Input file Identification group copied to {stats_file}")
@@ -144,7 +144,7 @@ def verify_gslc(
                 nisarqa.setup_report_pdf(product=product, report_pdf=report_pdf)
 
                 # Save frequency/polarization info to stats file
-                nisarqa.rslc.save_nisar_freq_metadata_to_h5(
+                nisarqa.save_nisar_freq_metadata_to_h5(
                     stats_h5=stats_h5, product=product
                 )
